partner/serializers.py
from rest_framework import serializers, viewsets

from .models import Partner


class PartnerSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Partner
        fields = (
            'name',
        )


class TaskViewSet(viewsets.ModelViewSet):
    queryset = Partner.objects.all()
    serializer_class = PartnerSerializer
