requirements.txt
wheel
Django==2.2
sqlparse==0.3.0                             # via Django
dj-database-url==0.5.0
django-admin-list-filter-dropdown==1.0.2
django-advanced-filters==1.1.0
django-braces==1.13.0                       # via django-advanced-filters
simplejson==3.16.0                          # via django-advanced-filters
django-adminfilters==1.1.0
djangorestframework==3.9.3
